pyproject.toml
[tool]
[tool.poetry]
name = "softpack-builder"
version = "0.1.0"
homepage = "https://github.com/wtsi-hgi/softpack-builder"
description = "SoftPack Builder provides services for building SoftPack environments."
authors = ["Altaf Ali <altaf.ali@sanger.ac.uk>"]
readme = "README.md"
license =  "MIT"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
packages = [
    { include = "softpack_builder" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
pip = "^23.0.1"
typer = {version = "^0.7.0"}
autoflake = {version = "^1.7.8", optional = true}
black = {version = "^23.1.0", optional = true}
bump2version = {version = "^1.0.1", optional = true}
essential-generators = {version = "^1.0", optional = true}
flake8 = {version = "^5.0.4", optional = true}
flake8-docstrings = {version = "^1.7.0", optional = true}
isort = {version = "^5.12.0", optional = true}
mkdocs = {version = "^1.4.2", optional = true}
mkdocs-autorefs = {version = "^0.4.1", optional = true}
mkdocs-include-markdown-plugin = {version = "^3.9.1", optional = true}
mkdocs-material = {version = "^9.1.0", optional = true}
mkdocstrings = {version = "^0.20.0", optional = true}
mkdocstrings-python = {version = "^0.8.3", optional = true}
mkdocs-material-extensions = {version = "^1.1.1", optional = true}
mkdocs-render-swagger-plugin = {version = "^0.0.4", optional = true}
mypy = {version = "^v1.0.1", optional = true}
mypy-extensions = {version = "^1.0.0", optional = true}
pytest = {version = "^7.2.2", optional = true}
pytest-asyncio = {version = "^0.21.0", optional = true}
pytest-cases = {version = "^3.6.14", optional = true}
pytest-cov = {version = "^4.0.0", optional = true}
pytest-httpx = {version = "^0.21.3", optional = true}
pre-commit = {version = "^3.1.1", optional = true}
toml = {version = "^0.10.2", optional = true}
tox = {version = "^4.4.6", optional = true}
twine = {version = "^4.0.2", optional = true}
virtualenv = {version = "^20.20.0", optional = true}
archspec = "^0.2.0"
bokeh = ">=2.4.3,<3.0"
fastapi = "^0.94.0"
greenlet = "^2.0.2"
prefect = "^2.8.4"
prefect-dask = "^0.2.3"
prefect-shell = "^0.1.5"
pydantic = "^1.10.6"
pygit2 = "^1.11.1"
python-box = "^7.0.1"
ruamel-yaml = "^0.17.21"
singleton-decorator = "^1.0.0"
sqlalchemy = "1.4.45"

[tool.poetry.extras]
test = [
    "autoflake",
    "black",
    "flake8",
    "flake8-docstrings",
    "isort",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "pytest-cases",
    "pytest-cov",
    "pytest-httpx"
    ]

dev = ["tox", "pre-commit", "virtualenv", "pip", "twine", "toml", "bump2version"]

doc = [
    "mkdocs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocstrings-python",
    "mkdocs-material-extension",
    "mkdocs-autorefs",
    "mkdocs-render-swagger-plugin"
    ]

[tool.poetry.scripts]
softpack-builder = 'softpack_builder.main:main'

[tool.black]
line-length = 79
skip-string-normalization = true
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:starlette"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
